reorganize, queue optional, queue with no_grad()
